Refactor code and switch Lisp to F#
The commit introduces a total refactor of the code in the day02 script. The original file which was written in Lisp (day02.lisp) was deleted and a new version of the script was added in F# (day02.fsx). Additionally, a slight code modification was done in day08.fsx script, simplifying a match expression.